requirements.txt
accelerate ==0.28.0
aiohttp ==3.9.3
aiosignal ==1.3.1
altair ==5.2.0
annotated-types ==0.6.0
anyio ==4.3.0
argon2-cffi ==23.1.0
argon2-cffi-bindings ==21.2.0
arrow ==1.3.0
asttokens ==2.4.1
async-lru ==2.0.4
async-timeout ==4.0.3
attrs ==23.2.0
Babel ==2.14.0
backcall ==0.2.0
beautifulsoup4 ==4.12.3
bitsandbytes ==0.43.0
bleach ==6.1.0
blinker ==1.7.0
bs4 ==0.0.2
cachetools ==5.3.3
certifi ==2024.2.2
cffi ==1.16.0
charset-normalizer ==3.3.2
click ==8.1.7
colorama ==0.4.6
comm ==0.2.2
dataclasses-json ==0.6.4
debugpy ==1.8.1
decorator ==5.1.1
defusedxml ==0.7.1
Deprecated ==1.2.14
dirtyjson ==1.0.8
distro ==1.9.0
docx ==0.2.4
docx2python ==2.8.0
docx2txt ==0.8
einops ==0.7.0
et-xmlfile ==1.1.0
exceptiongroup ==1.2.0
executing ==2.0.1
fastapi ==0.110.0
fastjsonschema ==2.19.1
filelock ==3.13.1
Flask ==3.0.3
fqdn ==1.5.1
frozenlist ==1.4.1
fsspec ==2024.3.1
gensim ==4.3.2
gitdb ==4.0.11
GitPython ==3.1.42
google-ai-generativelanguage ==0.4.0
google-api-core ==2.18.0
google-api-python-client ==2.125.0
google-auth ==2.29.0
google-auth-httplib2 ==0.2.0
google-generativeai ==0.4.1
googleapis-common-protos ==1.63.0
greenlet ==3.0.3
grpcio ==1.62.1
grpcio-status ==1.62.1
h11 ==0.14.0
httpcore ==1.0.4
httplib2 ==0.22.0
httpx ==0.27.0
huggingface-hub ==0.20.3
idna ==3.6
importlib_metadata ==7.1.0
ipykernel ==6.29.3
ipython ==8.18.1
ipywidgets ==8.1.2
isoduration ==20.11.0
itsdangerous ==2.2.0
jedi ==0.19.1
Jinja2 ==3.1.3
joblib ==1.3.2
json5 ==0.9.25
jsonpatch ==1.33
jsonpointer ==2.4
jsonschema ==4.21.1
jsonschema-specifications ==2023.12.1
jupyter ==1.0.0
jupyter_client ==8.6.1
jupyter-console ==6.6.3
jupyter_core ==5.7.2
jupyter-events ==0.10.0
jupyter-lsp ==2.2.5
jupyter_server ==2.14.0
jupyter_server_terminals ==0.5.3
jupyterlab ==4.1.6
jupyterlab_pygments ==0.3.0
jupyterlab_server ==2.26.0
jupyterlab_widgets ==3.0.10
langchain ==0.1.13
langchain-community ==0.0.29
langchain-core ==0.1.33
langchain-text-splitters ==0.0.1
langsmith ==0.1.31
llama-index ==0.10.29
llama-index-agent-openai ==0.2.2
llama-index-cli ==0.1.11
llama-index-core ==0.10.29
llama-index-embeddings-huggingface ==0.1.4
llama-index-embeddings-langchain ==0.1.2
llama-index-embeddings-openai ==0.1.7
llama-index-indices-managed-llama-cloud ==0.1.5
llama-index-legacy ==0.9.48
llama-index-llms-anyscale ==0.1.3
llama-index-llms-huggingface ==0.1.4
llama-index-llms-langchain ==0.1.3
llama-index-llms-ollama ==0.1.2
llama-index-llms-openai ==0.1.15
llama-index-llms-palm ==0.1.5
llama-index-llms-replicate ==0.1.3
llama-index-multi-modal-llms-openai ==0.1.5
llama-index-program-openai ==0.1.5
llama-index-question-gen-openai ==0.1.3
llama-index-readers-file ==0.1.11
llama-index-readers-llama-parse ==0.1.4
llama-parse ==0.4.1
llamaindex-py-client ==0.1.18
lxml ==5.1.0
markdown-it-py ==3.0.0
MarkupSafe ==2.1.5
marshmallow ==3.21.1
matplotlib-inline ==0.1.6
mdurl ==0.1.2
mistune ==3.0.2
mpmath ==1.3.0
multidict ==6.0.5
mypy-extensions ==1.0.0
nbclient ==0.10.0
nbconvert ==7.16.3
nbformat ==5.10.4
nest-asyncio ==1.6.0
networkx ==3.2.1
nltk ==3.8.1
notebook ==7.1.2
notebook_shim ==0.2.4
numpy ==1.26.4
openai ==1.14.2
openpyxl ==3.1.2
orjson ==3.9.15
overrides ==7.7.0
packaging ==23.2
pandas ==2.2.1
pandocfilters ==1.5.1
parso ==0.8.3
pexpect ==4.9.0
pickleshare ==0.7.5
pillow ==10.2.0
pip ==24.0
platformdirs ==4.2.0
prometheus_client ==0.20.0
prompt-toolkit ==3.0.43
proto-plus ==1.23.0
protobuf ==4.25.3
psutil ==5.9.8
ptyprocess ==0.7.0
pure-eval ==0.2.2
pyarrow ==15.0.2
pyasn1 ==0.6.0
pyasn1_modules ==0.4.0
pybind11 ==2.11.1
pycparser ==2.21
pydantic ==2.6.4
pydantic_core ==2.16.3
pydeck ==0.8.0
pydub ==0.25.1
Pygments ==2.17.2
PyMuPDF ==1.23.26
PyMuPDFb ==1.23.22
pyparsing ==3.1.2
pypdf ==4.1.0
python-dateutil ==2.9.0.post0
python-docx ==1.1.0
python-json-logger ==2.0.7
python-multipart ==0.0.9
pytz ==2024.1
pywin32 ==306
pywinpty ==2.0.13
PyYAML ==6.0.1
pyzmq ==25.1.2
qtconsole ==5.5.1
QtPy ==2.4.1
referencing ==0.34.0
regex ==2023.12.25
requests ==2.31.0
rfc3339-validator ==0.1.4
rfc3986-validator ==0.1.1
rich ==13.7.1
rpds-py ==0.18.0
rsa ==4.9
safetensors ==0.4.2
scikit-learn ==1.4.1.post1
scipy ==1.12.0
Send2Trash ==1.8.3
sentence-transformers ==2.5.1
setuptools ==49.2.1
six ==1.16.0
smart-open ==7.0.4
smmap ==5.0.1
sniffio ==1.3.1
soupsieve ==2.5
SQLAlchemy ==2.0.28
stack-data ==0.6.3
starlette ==0.36.3
streamlit ==1.32.2
striprtf ==0.0.26
sympy ==1.12
tenacity ==8.2.3
terminado ==0.18.1
threadpoolctl ==3.4.0
tiktoken ==0.6.0
tinycss2 ==1.2.1
tokenizers ==0.15.2
toml ==0.10.2
tomli ==2.0.1
toolz ==0.12.1
torch ==2.2.1
tornado ==6.4
tqdm ==4.66.2
traitlets ==5.14.2
transformers ==4.39.0
types-python-dateutil ==2.9.0.20240316
typing_extensions ==4.10.0
typing-inspect ==0.9.0
tzdata ==2024.1
tzlocal ==5.2
uri-template ==1.3.0
uritemplate ==4.1.1
urllib3 ==2.2.1
uvicorn ==0.29.0
validators ==0.23.2
watchdog ==4.0.0
wcwidth ==0.2.13
webcolors ==1.13
webencodings ==0.5.1
websocket-client ==1.7.0
Werkzeug ==3.0.2
widgetsnbextension ==4.0.10
wrapt ==1.16.0
yarl ==1.9.4
zipp ==3.18.1